Factor schema v1 impl into internal pkg
We want to add a schema v2 while still supporting existing schemav1
files.
