+++ internal/schemav1/schemav1.go
@@ -1,0 +1,1 @@
+package schemav1
